+++ src/components/day/Day.js
@@ -3,10 +3,16 @@ import _get from 'lodash/get';
 import _each from 'lodash/each';
 import _assign from 'lodash/assign';
 import moment from 'moment';
+import { getLocaleDateFormatInfo } from '../../utils';
 export class DayComponent extends BaseComponent {
   constructor(component, options, data) {
     super(component, options, data);
     this.validators.push('date');
+
+    const dateFormatInfo = getLocaleDateFormatInfo(options.language);
+    this.dayFirst = this.component.useLocaleSettings
+      ? dateFormatInfo.dayFirst
+      : this.component.dayFirst;
   }
 
   elementInfo() {
@@ -179,21 +185,17 @@ export class DayComponent extends BaseComponent {
       class: 'input-group row',
       style: 'width: 100%'
     });
-
     const subinputAtTheBottom = this.component.inputsLabelPosition === 'bottom';
-
-    const dayColumn = this.createDayInput(subinputAtTheBottom);
-    const monthColumn = this.createMonthInput(subinputAtTheBottom);
-    const yearColumn = this.createYearInput(subinputAtTheBottom);
+    const [ dayColumn, monthColumn, yearColumn ] = this.createInputs(subinputAtTheBottom);
 
     // Add the columns to the day select in the right order.
-    if (this.component.dayFirst && !_get(this.component, 'fields.day.hide', false)) {
+    if (this.dayFirst && !_get(this.component, 'fields.day.hide', false)) {
       inputGroup.appendChild(dayColumn)
     }
     if (!_get(this.component, 'fields.month.hide', false)) {
       inputGroup.appendChild(monthColumn);
     }
-    if (!this.component.dayFirst && !_get(this.component, 'fields.day.hide', false)) {
+    if (!this.dayFirst && !_get(this.component, 'fields.day.hide', false)) {
       inputGroup.appendChild(dayColumn);
     }
     if (!_get(this.component, 'fields.year.hide', false)) {
@@ -205,6 +207,19 @@ export class DayComponent extends BaseComponent {
     this.errorContainer = container;
     this.setInputStyles(inputGroup);
     container.appendChild(inputGroup);
+  }
+
+  createInputs(subinputAtTheBottom) {
+    return [
+      this.createDayInput(subinputAtTheBottom),
+      this.createMonthInput(subinputAtTheBottom),
+      this.createYearInput(subinputAtTheBottom),
+    ];
+  }
+
+  createViewOnlyInput() {
+    super.createViewOnlyInput();
+    this.createInputs();
   }
 
   setSubinputLabelStyle(label) {
@@ -319,5 +334,10 @@ export class DayComponent extends BaseComponent {
   getValueAt(index) {
     this.inputs[index].value = this.date.format(this.format);
     return this.inputs[index].value;
+  }
+
+  get view() {
+    const date = this.date;
+    return date.isValid() ? date.format(this.format) : null;
   }
 }