+++ src/utils/index.js
@@ -32,6 +32,22 @@ jsonLogic.add_operation('in', function(a, b) {
   return (b.indexOf(a) !== -1);
 });
 
+// Retrieve Any Date
+jsonLogic.add_operation("getDate", function(date){
+  return momentModule(date).toISOString()
+});
+
+// Set Relative Minimum Date
+jsonLogic.add_operation("relativeMinDate", function(relativeMinDate){
+  return momentModule().subtract(relativeMinDate, "days").toISOString()
+});
+
+// Set Relative Maximum Date
+jsonLogic.add_operation("relativeMaxDate", function(relativeMaxDate){
+  return momentModule().add(relativeMaxDate, "days").toISOString();
+});
+
+
 const FormioUtils = {
   jsonLogic, // Share
 
@@ -527,6 +543,17 @@ const FormioUtils = {
   },
   isValidDate(date) {
     return _isDate(date) && !_isNaN(date.getDate());
+  },
+  getLocaleDateFormatInfo(locale) {
+    const formatInfo = {};
+
+    const day = 21;
+    const exampleDate = new Date(2017, 11, day);
+    const localDateString = exampleDate.toLocaleDateString(locale);
+
+    formatInfo.dayFirst = localDateString.slice(0, 2) === day.toString();
+
+    return formatInfo;
   }
 };