+++ Changelog.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - CSS class ('radio-selected') for selected option of Radio component
 
+## 2.25.8
+### Added
+ - Support for JSONLogic dates.
+ - Added 'searchEnabled' option and defaulted it to false, user can enable it with component property
+ - Updated default value for 'removeItemButton' option to multi-select OR false, and if needed user can enable it with the component property
+ - Date formatting based on the locale configuration
+ - viewAsHTML feature.
+ - Confirmation dialog before a form/wizard is canceled.
+
 ## 2.25.7
 ### Fixed
  - Problem with the Select dropdown from re-rendering after it has been destroyed.